fix(clippy): correct doc comment indents in driver and compiler
Team Doc-Fix work — 8 warnings eliminated:
- driver/src/main.rs: 7x doc_overindented_list_items (40+ space indents reduced to 4)
- compiler/src/codegen/instr/body.rs: 1x doc_lazy_continuation in binop_result_type doc

Verified: cargo clippy -p simple-driver -W doc_overindented_list_items → 0
         cargo clippy -p simple-compiler -W doc_overindented_list_items -W doc_lazy_continuation → 0

+++ src/compiler_rust/compiler/src/codegen/cranelift.rs
@@ -254,7 +254,7 @@ fn reemit_clean_macho(malformed: &[u8]) -> Result<Vec<u8>, String> {
         let out_section = out.add_section(segment.as_bytes().to_vec(), name.as_bytes().to_vec(), kind);
         let data = section.data().unwrap_or(&[]);
         out.section_mut(out_section)
-            .set_data(data.to_vec(), section.align() as u64);
+            .set_data(data.to_vec(), section.align());
         section_map.insert(section.index(), out_section);
     }
